+++ packages/@expo/cli/src/start/server/metro/createServerComponentsMiddleware.ts
@@ -21,6 +21,7 @@ import {
   ExpoMetroOptions,
   getMetroOptionsFromUrl,
 } from '../middleware/metroOptions';
+import { getIpAddress } from '../../../utils/ip';
 
 const debug = require('debug')('expo:rsc') as typeof console.log;
 
@@ -58,6 +59,17 @@ export function createServerComponentsMiddleware(
     rscPath,
     onError: console.error,
     renderRsc: async (args) => {
+      // In development we should add simulated versions of common production headers.
+      if (args.headers['x-real-ip'] == null) {
+        args.headers['x-real-ip'] = getIpAddress();
+      }
+      if (args.headers['x-forwarded-for'] == null) {
+        args.headers['x-forwarded-for'] = args.headers['x-real-ip'];
+      }
+      if (args.headers['x-forwarded-proto'] == null) {
+        args.headers['x-forwarded-proto'] = 'http';
+      }
+
       // Dev server-only implementation.
       try {
         return await renderRscToReadableStream({
@@ -342,7 +354,7 @@ export function createServerComponentsMiddleware(
   async function renderRscToReadableStream(
     {
       input,
-      searchParams,
+      headers,
       method,
       platform,
       body,
@@ -352,7 +364,7 @@ export function createServerComponentsMiddleware(
       decodedBody,
     }: {
       input: string;
-      searchParams: URLSearchParams;
+      headers: Headers;
       method: 'POST' | 'GET';
       platform: string;
       body?: ReadableStream<Uint8Array>;
@@ -372,13 +384,17 @@ export function createServerComponentsMiddleware(
       assert(body, 'Server request must be provided when method is POST (server actions)');
     }
 
+    const context = getRscRenderContext(platform);
+
+    context['__expo_requestHeaders'] = headers;
+
     const { renderRsc } = await getRscRendererAsync(platform);
 
     return renderRsc(
       {
         body,
         decodedBody,
-        context: getRscRenderContext(platform),
+        context,
         config: {},
         input,
         contentType,
@@ -438,7 +454,7 @@ export function createServerComponentsMiddleware(
                 input,
                 method: 'GET',
                 platform,
-                searchParams: new URLSearchParams(),
+                headers: new Headers(),
                 ssrManifest,
               },
               true